// web-frontend/src/types/prediction.ts
export type PredictionResult = {
  id: number;
  Gender: string;
  Age: number;
  Driving_License: number;
  Region_Code: number;
  Previously_Insured: number;
  Vehicle_Age: string;
  Vehicle_Damage: string;
  Annual_Premium: number;
  Policy_Sales_Channel: number;
  Vintage: number;
  Response: number;
};